Scaled planets accordingly, set camera position, fixed two small bugs

+++ Simplex_17-18_Spring-master (1)/Simplex_17-18_Spring-master/00_Sandbox/Planet.cpp
@@ -46,7 +46,7 @@ void Planet::Move() {
 
 	force = mass * acceleration; 
 
-	newVelocity = velocity + ((force / mass) * CLOCKS_PER_SEC);
+	newVelocity = velocity + ((force / mass) * (float)CLOCKS_PER_SEC);
 
 	center += force;
